+++ assets/memberships-gate/editor.js
@@ -3,8 +3,7 @@
  * WordPress dependencies
  */
 import { sprintf, __ } from '@wordpress/i18n';
-import { withSelect, withDispatch } from '@wordpress/data';
-import { compose } from '@wordpress/compose';
+import { useSelect, useDispatch } from '@wordpress/data';
 import { Fragment, useEffect } from '@wordpress/element';
 import { Button, TextControl, CheckboxControl, SelectControl } from '@wordpress/components';
 import { PluginDocumentSettingPanel, PluginPostStatusInfo } from '@wordpress/edit-post';
@@ -34,7 +33,14 @@ const overlaySizes = [
 	{ value: 'full-width', label: __( 'Full Width', 'newspack' ) },
 ];
 
-const GateEdit = ( { editPost, meta } ) => {
+function GateEdit() {
+	const { meta } = useSelect( select => {
+		const { getEditedPostAttribute } = select( 'core/editor' );
+		return {
+			meta: getEditedPostAttribute( 'meta' ),
+		};
+	} );
+	const { editPost } = useDispatch( 'core/editor' );
 	useEffect( () => {
 		const wrapper = document.querySelector( '.editor-styles-wrapper' );
 		if ( ! wrapper ) {
@@ -115,6 +121,7 @@ const GateEdit = ( { editPost, meta } ) => {
 			>
 				<TextControl
 					type="number"
+					min="0"
 					value={ meta.visible_paragraphs }
 					label={ __( 'Default paragraph count', 'newspack' ) }
 					onChange={ value => editPost( { meta: { visible_paragraphs: value } } ) }
@@ -134,23 +141,65 @@ const GateEdit = ( { editPost, meta } ) => {
 					) }
 				/>
 			</PluginDocumentSettingPanel>
+			<PluginDocumentSettingPanel
+				name="memberships-gate-metering-panel"
+				title={ __( 'Metering', 'newspack' ) }
+			>
+				<CheckboxControl
+					label={ __( 'Enable metering', 'newspack' ) }
+					checked={ meta.metering }
+					onChange={ value => editPost( { meta: { metering: value } } ) }
+					help={ __(
+						'Implement metering to configure access to restricted content before showing the gate.',
+						'newspack'
+					) }
+				/>
+				{ meta.metering && (
+					<Fragment>
+						<TextControl
+							type="number"
+							min="0"
+							value={ meta.metering_anonymous_count }
+							label={ __( 'Available views for anonymous readers', 'newspack' ) }
+							onChange={ value => editPost( { meta: { metering_anonymous_count: value } } ) }
+							help={ __(
+								'Number of times an anonymous reader can view gated content. If set to 0, anonymous readers will always render the gate.',
+								'newspack'
+							) }
+						/>
+						<TextControl
+							type="number"
+							min="0"
+							value={ meta.metering_registered_count }
+							label={ __( 'Available views for registered readers', 'newspack' ) }
+							onChange={ value => editPost( { meta: { metering_registered_count: value } } ) }
+							help={ __(
+								'Number of times a registered reader can view gated content. If set to 0, registered readers without membership plan will always render the gate.',
+								'newspack'
+							) }
+						/>
+						<SelectControl
+							label={ __( 'Time period', 'newspack' ) }
+							value={ meta.metering_period }
+							options={ [
+								{ value: 'day', label: __( 'Day', 'newspack' ) },
+								{ value: 'week', label: __( 'Week', 'newspack' ) },
+								{ value: 'month', label: __( 'Month', 'newspack' ) },
+							] }
+							onChange={ value => editPost( { meta: { metering_period: value } } ) }
+							help={ __(
+								'The time period during which the metering views will be counted. For example, if the metering period is set to a week, the metering views will be reset every week.',
+								'newspack'
+							) }
+						/>
+					</Fragment>
+				) }
+			</PluginDocumentSettingPanel>
 		</Fragment>
 	);
-};
-
-const GateEditWithSelect = compose( [
-	withSelect( select => {
-		const { getEditedPostAttribute } = select( 'core/editor' );
-		const meta = getEditedPostAttribute( 'meta' );
-		return { meta };
-	} ),
-	withDispatch( dispatch => {
-		const { editPost } = dispatch( 'core/editor' );
-		return { editPost };
-	} ),
-] )( GateEdit );
+}
 
 registerPlugin( 'newspack-memberships-gate', {
-	render: GateEditWithSelect,
+	render: GateEdit,
 	icon: null,
 } );