+++ includes/templates/block-patterns/homepage/06.php
@@ -6,8 +6,7 @@
  */
 
 return array(
-	'content' => '<!-- wp:group {"className":"newspack-pattern homepage-posts__style-26"} -->
-<div class="wp-block-group newspack-pattern homepage-posts__style-26"><div class="wp-block-group__inner-container"><!-- wp:columns {"className":"is-style-first-col-to-second"} -->
+	'content' => '<!-- wp:columns {"className":"is-style-first-col-to-second"} -->
 <div class="wp-block-columns is-style-first-col-to-second"><!-- wp:column {"width":"50%"} -->
 <div class="wp-block-column" style="flex-basis:50%"><!-- wp:newspack-blocks/homepage-articles {"showAuthor":false,"postsToShow":1,"typeScale":6} /-->
 
@@ -21,41 +20,17 @@ return array(
 <!-- wp:column {"width":"25%"} -->
 <div class="wp-block-column" style="flex-basis:25%"><!-- wp:newspack-blocks/homepage-articles {"className":"is-style-borders","excerptLength":12,"showImage":false,"showAuthor":false,"showAvatar":false,"postsToShow":7,"mediaPosition":"left","typeScale":2,"imageScale":1} /--></div>
 <!-- /wp:column --></div>
-<!-- /wp:columns --></div></div>
-<!-- /wp:group -->
+<!-- /wp:columns -->
 
 ' . $donate_block . '
 
-<!-- wp:group {"className":"newspack-pattern homepage-posts__style-28"} -->
-<div class="wp-block-group newspack-pattern homepage-posts__style-28"><div class="wp-block-group__inner-container"><!-- wp:columns -->
-<div class="wp-block-columns"><!-- wp:column {"width":"20%"} -->
-<div class="wp-block-column" style="flex-basis:20%"><!-- wp:newspack-blocks/homepage-articles {"showExcerpt":false,"showAuthor":false,"postsToShow":1,"authors":[],"categories":[],"tags":[],"tagExclusions":[],"specificPosts":[],"typeScale":2} /--></div>
-<!-- /wp:column -->
-
-<!-- wp:column {"width":"20%"} -->
-<div class="wp-block-column" style="flex-basis:20%"><!-- wp:newspack-blocks/homepage-articles {"showExcerpt":false,"showAuthor":false,"postsToShow":1,"authors":[],"categories":[],"tags":[],"tagExclusions":[],"specificPosts":[],"typeScale":2} /--></div>
-<!-- /wp:column -->
-
-<!-- wp:column {"width":"20%"} -->
-<div class="wp-block-column" style="flex-basis:20%"><!-- wp:newspack-blocks/homepage-articles {"showExcerpt":false,"showAuthor":false,"postsToShow":1,"authors":[],"categories":[],"tags":[],"tagExclusions":[],"specificPosts":[],"typeScale":2} /--></div>
-<!-- /wp:column -->
-
-<!-- wp:column {"width":"20%"} -->
-<div class="wp-block-column" style="flex-basis:20%"><!-- wp:newspack-blocks/homepage-articles {"showExcerpt":false,"showAuthor":false,"postsToShow":1,"authors":[],"categories":[],"tags":[],"tagExclusions":[],"specificPosts":[],"typeScale":2} /--></div>
-<!-- /wp:column -->
-
-<!-- wp:column {"width":"20%"} -->
-<div class="wp-block-column" style="flex-basis:20%"><!-- wp:newspack-blocks/homepage-articles {"showExcerpt":false,"showAuthor":false,"postsToShow":1,"authors":[],"categories":[],"tags":[],"tagExclusions":[],"specificPosts":[],"typeScale":2} /--></div>
-<!-- /wp:column --></div>
-<!-- /wp:columns --></div></div>
-<!-- /wp:group -->
+<!-- wp:newspack-blocks/homepage-articles {"showExcerpt":false,"showAuthor":false,"postLayout":"grid","columns":5,"postsToShow":5,"authors":[],"categories":[],"tags":[],"tagExclusions":[],"specificPosts":[],"typeScale":2} /-->
 
 <!-- wp:separator {"className":"is-style-wide"} -->
 <hr class="wp-block-separator is-style-wide"/>
 <!-- /wp:separator -->
 
-<!-- wp:group {"className":"newspack-pattern homepage-posts__style-29"} -->
-<div class="wp-block-group newspack-pattern homepage-posts__style-29"><div class="wp-block-group__inner-container"><!-- wp:columns -->
+<!-- wp:columns -->
 <div class="wp-block-columns"><!-- wp:column {"width":"50%"} -->
 <div class="wp-block-column" style="flex-basis:50%"><!-- wp:newspack-blocks/homepage-articles {"excerptLength":25,"showAuthor":false,"showAvatar":false,"mediaPosition":"left","typeScale":2,"imageScale":2} /--></div>
 <!-- /wp:column -->
@@ -67,7 +42,6 @@ return array(
 <!-- wp:column {"width":"25%"} -->
 <div class="wp-block-column" style="flex-basis:25%"><!-- wp:newspack-blocks/homepage-articles {"className":"is-style-borders","showExcerpt":false,"showImage":false,"showAuthor":false,"showAvatar":false,"postsToShow":5,"typeScale":2} /--></div>
 <!-- /wp:column --></div>
-<!-- /wp:columns --></div></div>
-<!-- /wp:group -->',
+<!-- /wp:columns -->',
 	'image'   => '<svg viewBox="0 0 142 102" xmlns="http://www.w3.org/2000/svg"><path d="m37.4722 0h67.0558v50.2917h-67.0558z"/><path d="m108.472.369792h31.063v2.342018h-31.063z"/><path d="m108.472 18.4896h22.188v2.342h-22.188z"/><path d="m108.472 36.6094h31.063v2.342h-31.063z"/><path d="m108.472 54.7292h24.16v2.342h-24.16z"/><path d="m108.472 72.849h31.063v2.342h-31.063z"/><path d="m108.472 96.3924h33.528v2.342h-33.528z"/><path d="m108.472 75.4375h21.941v2.342h-21.941z"/><path d="m108.472 98.9809h4.191v2.3421h-4.191z"/><path d="m108.472 4.06771h33.528v1.84896h-33.528z"/><path d="m108.472 22.1875h30.2v1.849h-30.2z"/><path d="m108.472 40.3073h28.474v1.8489h-28.474z"/><path d="m108.472 58.4271h31.802v1.8489h-31.802z"/><path d="m108.472 81.9705h29.584v1.8489h-29.584z"/><path d="m108.472 79.2587h33.035v1.8489h-33.035z"/><path d="m108.472 6.77951h27.242v1.84896h-27.242z"/><path d="m108.472 24.8993h17.627v1.849h-17.627z"/><path d="m108.472 43.0191h17.627v1.849h-17.627z"/><path d="m108.472 61.1389h16.518v1.8489h-16.518z"/><path d="m108.472 84.6823h33.035v1.8489h-33.035z"/><path d="m108.472 10.3542h9.245v1.8489h-9.245z"/><path d="m108.472 28.474h9.245v1.8489h-9.245z"/><path d="m108.472 46.5938h9.245v1.8489h-9.245z"/><path d="m108.472 64.7135h9.245v1.849h-9.245z"/><path d="m108.472 88.2569h9.245v1.849h-9.245z"/><path d="m0 0h33.5278v25.1458h-33.5278z"/><path d="m0 37.4722h33.5278v25.1459h-33.5278z"/><path d="m0 74.9444h33.5278v25.1456h-33.5278z"/><path d="m0 27.1181h27.8576v3.0816h-27.8576z"/><path d="m0 64.5903h30.3229v3.0816h-30.3229z"/><path d="m0 31.6788h9.24479v1.849h-9.24479z"/><path d="m0 69.151h9.24479v1.849h-9.24479z"/><path d="m37.4722 53.8663h67.0558v6.1632h-67.0558z"/><path d="m37.4722 60.8924h35.7466v6.1632h-35.7466z"/><path d="m37.4722 70.1372h67.0558v2.4652h-67.0558z"/><path d="m37.4722 74.0816h30.6927v2.4653h-30.6927z"/><path d="m37.4722 79.0122h10.3542v1.9722h-10.3542z"/><path d="m61.2622 86.901h27.6111v2.3421h-27.6111z"/><path d="m61.2622 90.599h43.2658v1.8489h-43.2658z"/><path d="m61.2622 93.3108h43.2658v1.8489h-43.2658z"/><path d="m61.2622 96.0226h21.4479v1.8489h-21.4479z"/><path d="m61.2622 99.5972h9.2447v1.8488h-9.2447z"/><path d="m37.4722 86.5312h21.3247v15.4688h-21.3247z"/></svg>',
 );