+++ webpack.config.js
@@ -76,6 +76,18 @@ const webpackConfig = getBaseWebpackConfig(
 				'memberships-gate',
 				'overlay.js'
 			),
+			'memberships-gate-metering': path.join(
+				__dirname,
+				'assets',
+				'memberships-gate',
+				'metering.js'
+			),
+			'memberships-gate-block-patterns': path.join(
+				__dirname,
+				'assets',
+				'memberships-gate',
+				'block-patterns.js'
+			),
 		},
 	}
 );